✨ certif: use PixTable on invitations list
