--- certif/app/components/team/invitations-list-item.hbs
@@ -1,40 +1,0 @@
-<tr aria-label="{{t 'pages.team-invitations.table.row.aria-label'}}">
-  <td>{{@invitation.email}}</td>
-  <td>
-    {{dayjs-format @invitation.updatedAt "DD/MM/YYYY [-] HH:mm"}}
-  </td>
-  <td>
-    <div class="invitations-list-item__actions">
-      <PixTooltip @isInline={{true}}>
-        <:triggerElement>
-          <PixIconButton
-            @ariaLabel={{t "pages.team-invitations.actions.resend-invitation"}}
-            @iconName="refresh"
-            @triggerAction={{fn @onResendInvitationButtonClicked @invitation}}
-            @withBackground={{true}}
-            disabled={{@invitation.isResendingInvitation}}
-            aria-disabled={{@invitation.isResendingInvitation}}
-          />
-        </:triggerElement>
-        <:tooltip>
-          {{t "pages.team-invitations.actions.resend-invitation"}}
-        </:tooltip>
-      </PixTooltip>
-
-      <PixTooltip @isInline={{true}}>
-        <:triggerElement>
-          <PixIconButton
-            @ariaLabel={{t "pages.team-invitations.actions.cancel-invitation"}}
-            @iconName="delete"
-            @plainIcon={{true}}
-            @triggerAction={{fn @onCancelInvitationButtonClicked @invitation}}
-            @withBackground={{true}}
-          />
-        </:triggerElement>
-        <:tooltip>
-          {{t "pages.team-invitations.actions.cancel-invitation"}}
-        </:tooltip>
-      </PixTooltip>
-    </div>
-  </td>
-</tr>
