+++ certif/app/components/team/invitations-list.hbs
@@ -1,20 +1,59 @@
-<div class="panel table content-text--small">
-  <table>
-    <thead>
-      <tr>
-        <Table::Header>{{t "pages.team-invitations.table.labels.email-address"}}</Table::Header>
-        <Table::Header>{{t "pages.team-invitations.table.labels.last-sending-date"}}</Table::Header>
-        <Table::Header @size="small">{{t "pages.team-invitations.table.labels.actions"}}</Table::Header>
-      </tr>
-    </thead>
-    <tbody>
-      {{#each @invitations as |invitation|}}
-        <Team::InvitationsListItem
-          @invitation={{invitation}}
-          @onCancelInvitationButtonClicked={{@onCancelInvitationButtonClicked}}
-          @onResendInvitationButtonClicked={{@onResendInvitationButtonClicked}}
-        />
-      {{/each}}
-    </tbody>
-  </table>
-</div>
+<PixTable @data={{@invitations}} @variant="certif" @caption={{t "pages.team-invitations.table.caption"}}>
+  <:columns as |invitation context|>
+    <PixTableColumn @context={{context}}>
+      <:header>
+        {{t "pages.team-invitations.table.labels.email-address"}}
+      </:header>
+      <:cell>
+        {{invitation.email}}
+      </:cell>
+    </PixTableColumn>
+    <PixTableColumn @context={{context}}>
+      <:header>
+        {{t "pages.team-invitations.table.labels.last-sending-date"}}
+      </:header>
+      <:cell>
+        {{dayjs-format invitation.updatedAt "DD/MM/YYYY [-] HH:mm"}}
+      </:cell>
+    </PixTableColumn>
+    <PixTableColumn @context={{context}}>
+      <:header>
+        {{t "pages.team-invitations.table.labels.actions"}}
+      </:header>
+      <:cell>
+        <div class="invitations-list__actions">
+          <PixTooltip @isInline={{true}}>
+            <:triggerElement>
+              <PixIconButton
+                @ariaLabel={{t "pages.team-invitations.actions.resend-invitation"}}
+                @iconName="refresh"
+                @triggerAction={{fn @onResendInvitationButtonClicked invitation}}
+                @withBackground={{true}}
+                disabled={{invitation.isResendingInvitation}}
+                aria-disabled={{invitation.isResendingInvitation}}
+              />
+            </:triggerElement>
+            <:tooltip>
+              {{t "pages.team-invitations.actions.resend-invitation"}}
+            </:tooltip>
+          </PixTooltip>
+
+          <PixTooltip @isInline={{true}}>
+            <:triggerElement>
+              <PixIconButton
+                @ariaLabel={{t "pages.team-invitations.actions.cancel-invitation"}}
+                @iconName="delete"
+                @plainIcon={{true}}
+                @triggerAction={{fn @onCancelInvitationButtonClicked invitation}}
+                @withBackground={{true}}
+              />
+            </:triggerElement>
+            <:tooltip>
+              {{t "pages.team-invitations.actions.cancel-invitation"}}
+            </:tooltip>
+          </PixTooltip>
+        </div>
+      </:cell>
+    </PixTableColumn>
+  </:columns>
+</PixTable>
