+++ certif/app/styles/app.scss
@@ -50,7 +50,7 @@
 @use 'components/session-supervising/candidate-list' as *;
 @use 'components/session-supervising/handle-live-alert-modal' as *;
 @use 'components/session-supervising/live-alert-handled-modal' as *;
-@use 'components/team/invitations-list-item' as *;
+@use 'components/team/invitations-list' as *;
 @use 'components/dropdown' as *;
 @use 'flatpickr' as *;
