+++ certif/app/styles/components/team/invitations-list.scss
@@ -1,5 +1,4 @@
-.invitations-list-item__actions {
+.invitations-list__actions {
   display: flex;
   gap: var(--pix-spacing-6x);
-  width: 38px;
 }
